Use viewBoxF for SVG cacheAsImage and cacheAsPicture
- Fix QPicture SVG symbols not scaling with Composer page zoom
- Support non-squared SVG in output to print/image
- Make SVG symbol output to print/image more accurate by setting 'size' to double (casting to int only for QImage)

+++ src/core/symbology-ng/qgsmarkersymbollayerv2.cpp
@@ -643,8 +643,8 @@ void QgsSvgMarkerSymbolLayerV2::renderPoint( const QPointF& point, QgsSymbolV2Re
     outputOffset = _rotatedOffset( outputOffset, mAngle );
   p->translate( point + outputOffset );
 
-  int size = ( int )( context.outputLineWidth( mSize ) );
-  if ( size < 1 ) //don't render symbols with size below one pixel
+  double size = context.outputLineWidth( mSize );
+  if ( size < 1.0 ) //don't render symbols with size below one pixel
   {
     return;
   }
